--- polls/tests.py
@@ -1,189 +1,0 @@
-import datetime
-
-from django.urls import reverse
-from django.test import TestCase
-from django.utils import timezone
-
-from .models import Question
-
-
-class QuestionModelTests(TestCase):
-    def test_question_with_future_pub_date(self):
-        """Question with future pub date should not be shown"""
-        future_question = Question.objects.create(
-            question_text="Future Question",
-            pub_date=timezone.now() + timezone.timedelta(days=1)
-        )
-        self.assertFalse(future_question.is_published())
-
-    def test_question_with_past_pub_date(self):
-        """Question with pub date in the past should be published."""
-        past_question = Question.objects.create(
-            question_text="Past Question",
-            pub_date=timezone.now() - timezone.timedelta(days=1)
-        )
-        self.assertTrue(past_question.is_published())
-
-    def test_question_with_default_pub_date(self):
-        """Question with the default pub date should be published now."""
-        present_question = Question.objects.create(
-            question_text="Present Question",
-            pub_date=timezone.now()
-        )
-        self.assertTrue(present_question.is_published())
-
-    def test_cannot_vote_after_end_date(self):
-        """Cannot vote if the end_date is in the past."""
-        now = timezone.now()
-        question_with_end_date = Question.objects.create(
-            question_text="Question with End Date",
-            pub_date=now - timezone.timedelta(days=2),
-            end_date=now - timezone.timedelta(days=1)
-        )
-        self.assertFalse(question_with_end_date.can_vote())
-
-    def test_can_vote_within_voting_period(self):
-        """Can vote if the current time is within the voting period."""
-        question_with_open_voting = Question.objects.create(
-            question_text="Question with no end date",
-            pub_date=timezone.now() - timezone.timedelta(days=1))
-        self.assertTrue(question_with_open_voting.can_vote())
-
-    def test_cannot_vote_if_not_published(self):
-        """Cannot vote if the question is not yet published."""
-        future_question = Question.objects.create(
-            question_text="Future Question",
-            pub_date=timezone.now() + timezone.timedelta(days=1)
-        )
-        self.assertFalse(future_question.can_vote())
-
-    def test_can_vote_with_future_end_date(self):
-        """
-        Can vote if the current time is within the voting period
-        and the end_date is in the future.
-        """
-        now = timezone.now()
-        question_with_future_end_date = Question.objects.create(
-            question_text="Question with Future End Date",
-            pub_date=now - timezone.timedelta(days=2),
-            end_date=now + timezone.timedelta(days=1)
-        )
-        self.assertTrue(question_with_future_end_date.can_vote())
-
-    def test_was_published_recently_with_future_question(self):
-        """
-        was_published_recently() returns False for questions whose pub_date
-        is in the future.
-        """
-        time = timezone.now() + datetime.timedelta(days=30)
-        future_question = Question(pub_date=time)
-        self.assertIs(future_question.was_published_recently(), False)
-
-    def test_was_published_recently_with_old_question(self):
-        """
-        was_published_recently() returns False for questions whose pub_date
-        is older than 1 day.
-        """
-        time = timezone.now() - datetime.timedelta(days=1, seconds=1)
-        old_question = Question(pub_date=time)
-        self.assertIs(old_question.was_published_recently(), False)
-
-    def test_was_published_recently_with_recent_question(self):
-        """
-        was_published_recently() returns True for questions whose pub_date
-        is within the last day.
-        """
-        time = (timezone.now() -
-                datetime.timedelta(hours=23, minutes=59, seconds=59))
-        recent_question = Question(pub_date=time)
-        self.assertIs(recent_question.was_published_recently(), True)
-
-
-def create_question(question_text, days):
-    """
-    Create a question with the given `question_text` and published the
-    given number of `days` offset to now (negative for questions published
-    in the past, positive for questions that have yet to be published).
-    """
-    time = timezone.now() + timezone.timedelta(days=days)
-    return Question.objects.create(question_text=question_text, pub_date=time)
-
-
-class QuestionIndexViewTests(TestCase):
-    def test_no_questions(self):
-        """
-        If no questions exist, an appropriate message is displayed.
-        """
-        response = self.client.get(reverse("polls:index"))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerySetEqual(response.context["latest_question_list"], [])
-
-    def test_past_question(self):
-        """
-        Questions with a pub_date in the past are displayed on the
-        index page.
-        """
-        question = create_question(question_text="Past question.", days=-30)
-        response = self.client.get(reverse("polls:index"))
-        self.assertQuerySetEqual(
-            response.context["latest_question_list"],
-            [question],
-        )
-
-    def test_future_question(self):
-        """
-        Questions with a pub_date in the future aren't displayed on
-        the index page.
-        """
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse("polls:index"))
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerySetEqual(response.context["latest_question_list"], [])
-
-    def test_future_question_and_past_question(self):
-        """
-        Even if both past and future questions exist, only past questions
-        are displayed.
-        """
-        question = create_question(question_text="Past question.", days=-30)
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse("polls:index"))
-        self.assertQuerySetEqual(
-            response.context["latest_question_list"],
-            [question],
-        )
-
-    def test_two_past_questions(self):
-        """The questions index page may display multiple questions."""
-        question1 = create_question(question_text="Past question 1.", days=-30)
-        question2 = create_question(question_text="Past question 2.", days=-5)
-        response = self.client.get(reverse("polls:index"))
-        self.assertQuerySetEqual(
-            response.context["latest_question_list"],
-            [question2, question1],
-        )
-
-
-class QuestionDetailViewTests(TestCase):
-    def test_future_question(self):
-        """
-        The detail view of a question with a pub_date in the future
-        returns a 404 not found.
-        """
-        future_question = create_question(question_text='Future question.',
-                                          days=5)
-        url = reverse('polls:detail', args=(future_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-
-    def test_past_question(self):
-        """
-        The detail view of a question with a pub_date in the past
-        displays the question's text.
-        """
-        past_question = create_question(question_text='Past Question.',
-                                        days=-5)
-        url = reverse('polls:detail', args=(past_question.id,))
-        response = self.client.get(url)
-        self.assertContains(response, past_question.question_text)
