+++ polls/tests/test_auth.py
@@ -1,0 +1,7 @@
+from django.urls import reverse
+from django.test import TestCase
+from .utils import create_question
+
+
+class UserAuthTests(TestCase):
+    pass
